+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * [#1645](https://github.com/crypto-org-chain/cronos/pull/1645) Gen test tx in parallel even in single node.
 * (testground)[#1644](https://github.com/crypto-org-chain/cronos/pull/1644) load generator retry with backoff on error.
 * [#1648](https://github.com/crypto-org-chain/cronos/pull/1648) Add abort OE in PrepareProposal.
+* (testground)[#]() Benchmark use cosmos broadcast rpc.
 
 *Oct 14, 2024*